rename project, packeges
