+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.example</groupId>
-    <artifactId>Home_Task_1</artifactId>
+    <artifactId>Avic</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
