+++ src/test/java/com/epam/ta/avic/ui/pages/LeftSideMenuTest.java
@@ -1,10 +1,10 @@
-package com.epam.ta.hometask001.ui.pages;
+package com.epam.ta.avic.ui.pages;
 
-import com.epam.ta.hometask001.ui.runner.TestRunner;
+import com.epam.ta.avic.ui.runner.TestRunner;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import static com.epam.ta.hometask001.ui.locators.HomePageLocators.*;
+import static com.epam.ta.avic.ui.locators.HomePageLocators.*;
 
 public class LeftSideMenuTest extends TestRunner {
